backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai 
pymupdf
python-multipart
pymongo
faiss-cpu
requests
beautifulsoup4
python-docx
ollama
python-dotenv # Added for loading .env files